Move TestStateCompiler to state package and clean up
